% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.r
\name{cpa}
\alias{cpa}
\title{Cutting Plane Algorithm}
\usage{
cpa(x, y, ...)
}
\arguments{
\item{x}{n x d matrix of features}

\item{y}{n x 1 matrix of classification labels {-1, 1}}
}
\description{
Cutting Plane Algorithm
}
