% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.r
\name{eval_auc}
\alias{eval_auc}
\alias{eval_cal}
\alias{eval_precision}
\alias{eval_recall}
\alias{eval_accuracy}
\alias{eval_logloss}
\title{Evaluation Metrics}
\usage{
eval_auc(model_fit, new_data, ...)

eval_cal(model_fit, new_data, grouped = FALSE, ...)

eval_precision(model_fit, new_data, threshold = 0.5, ...)

eval_recall(model_fit, new_data, threshold = 0.5, ...)

eval_accuracy(model_fit, new_data, threshold = 0.5, ...)

eval_logloss(model_fit, new_data)
}
\arguments{
\item{model_fit}{model fit from \code{lcpa}}

\item{new_data}{data frame}

\item{threshold}{vector of values to use as threshold cutoffs}
}
\description{
Evaluation Metrics
}
