% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpa.r
\name{lcpa}
\alias{lcpa}
\alias{lcpa.default}
\alias{lcpa.data.frame}
\title{Lattice Cutting Plane Algorithm}
\usage{
lcpa(x, ...)

\method{lcpa}{default}(x, ...)

\method{lcpa}{data.frame}(
  df,
  formula,
  weights = NULL,
  R_max = NULL,
  time_limit = 60,
  logfile = random_logfile(),
  engine = c("cplex", "glpk"),
  ...
)
}
\description{
Lattice Cutting Plane Algorithm
}
