% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.r, R/predict.r, R/recipes.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glance}
\alias{augment}
\alias{tidy}
\alias{predict}
\alias{prep}
\alias{bake}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{augment}}, \code{\link[generics]{tidy}}}

  \item{recipes}{\code{\link[recipes]{prep}}, \code{\link[recipes]{bake}}}

  \item{stats}{\code{\link[stats]{predict}}}
}}

