% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskslimr-package.R
\docType{package}
\name{riskslimr-package}
\alias{riskslimr}
\alias{riskslimr-package}
\title{riskslimr: Implementation of RiskSLIM model in R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zamorarr/riskslimr}
  \item Report bugs at \url{https://github.com/zamorarr/riskslimr/issues}
}

}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
